+++ CHANGELOG.md
@@ -33,6 +33,7 @@ https://github.com/nwnxee/unified/compare/build8193.34...HEAD
 ### Changed
 - Events: added event data `VERSION_MAJOR`/`VERSION_MINOR`/`PLATFORM_ID` to `NWNX_ON_CLIENT_CONNECT_*`
 - Object: DoSpellLevelAbsorption() can now override the spellId, the level and the school of the casted spell
+- Object: DoSpellImmunity() can now override the spellId
 
 ### Deprecated
 - N/A